client api

+++ src/client/client.service.ts
@@ -1,6 +1,6 @@
 import { ForbiddenException, Injectable } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
-import { CreateClientDto } from './dto';
+import { CreateClientDto, UpdateClientDto } from './dto';
 import { PrismaClientKnownRequestError } from '@prisma/client/runtime/library';
 
 @Injectable()
@@ -12,7 +12,11 @@ export class ClientService {
         id: clientId,
       },
     });
-    return client;
+    if (client) {
+      return client;
+    } else {
+      throw new ForbiddenException('No client found');
+    }
   }
 
   async createClient(dto: CreateClientDto) {
@@ -34,30 +38,50 @@ export class ClientService {
     }
   }
 
-  async updateClient(clientId: number, dto) {
-    const updatedClient = await this.prisma.client.update({
-      where: {
-        id: clientId,
-      },
-      data: {
-        ...dto,
-      },
-    });
-
-    return updatedClient;
+  async updateClient(clientId: number, dto: UpdateClientDto) {
+    try {
+      const updatedClient = await this.prisma.client.update({
+        where: {
+          id: clientId,
+        },
+        data: {
+          ...dto,
+        },
+      });
+      return updatedClient;
+    } catch (error) {
+      throw error;
+    }
   }
 
   async deleteClient(clientId: number) {
-    const deletedClient = await this.prisma.client.delete({
-      where: {
-        id: clientId,
-      },
-    });
-    return deletedClient;
+    try {
+      const client = await this.prisma.client.findUnique({
+        where: {
+          id: clientId,
+        },
+      });
+      if (client) {
+        const deletedClient = await this.prisma.client.delete({
+          where: {
+            id: clientId,
+          },
+        });
+        return deletedClient;
+      } else {
+        throw new ForbiddenException('No client found');
+      }
+    } catch (error) {
+      throw error;
+    }
   }
 
   async getAllClients() {
-    const clients = await this.prisma.client.findMany();
-    return clients;
+    try {
+      const clients = await this.prisma.client.findMany();
+      return clients;
+    } catch (error) {
+      throw error;
+    }
   }
 }
